+++ server/accountHolder.h
@@ -15,13 +15,13 @@ public:
     const char* getFullName();                //Returns the full name of the account holder
     const char* getAddress();                 //Returns the address of the account holder
     const char* getSSN();                     //Returns the SSN of the account holder
-    double getCreditScore();                  //Returns the credit score of the account holder
-    double getAccount();                      //Returns the account number associated with this holder
-    void addAccount(double accountNumber);    //Adds an account associated with the account holder
+    long getCreditScore();                  //Returns the credit score of the account holder
+    long getAccount();                      //Returns the account number associated with this holder
+    void addAccount(long accountNumber);    //Adds an account associated with the account holder
     //Set Personal information functions
     void setAddress(const char* addressIn);
     void setSSN(const char* SSNIn);
-    void setCreditScore(double ScoreIn);
+    void setCreditScore(long ScoreIn);
 
 private:
     std::string name;
@@ -29,8 +29,8 @@ private:
     std::string lastName;
     std::string address;
     std::string SSN;
-    double score;
-    double account;
+    long score;
+    long account;
 
 };
