+++ server/log.h
@@ -12,8 +12,8 @@ using namespace std;
 class log{
 public:
     static bool createFile(string user, string password);
-    static void writeFile(string user, string password, double balance, double amount, int flag);
-    static double getBalance(string user, string password);
+    static void writeFile(string user, string password, long balance, long amount, int flag);
+    static long getBalance(string user, string password);
     static string getPassword(string user, string password);
     static void openLog(string user, string password);
 };
